[Refactor] P2P split Phase 1: move master-side P2P files into p2p/master
Pure relocation, zero logic change (Phase 1 commit 1/4):
- git mv p2p_master_{service,client,metric_manager}, p2p_segment_manager,
  p2p_rpc_{service,types} into mooncake-store/{include,src}/p2p/master/
- rewrite all #include paths accordingly
- introduce mooncake_store_p2p CMake target holding src/p2p/**; it links
  mooncake_store wholesale until base-class inheritance is severed (Phase 2);
  leaf-only layering lands in Phase 3
- link the new lib into mooncake_master, unit tests and the python module
