+++ mooncake-integration/CMakeLists.txt
@@ -90,7 +90,7 @@ if (WITH_STORE)
             transfer_engine
             glog::glog
             gflags::gflags
-            mooncake_store
+            $<LINK_GROUP:RESCAN,mooncake_store,mooncake_store_p2p>
             cachelib_memory_allocator
         )
     else()
@@ -98,7 +98,7 @@ if (WITH_STORE)
             transfer_engine
             glog::glog
             gflags::gflags
-            mooncake_store
+            $<LINK_GROUP:RESCAN,mooncake_store,mooncake_store_p2p>
             cachelib_memory_allocator
         )
     endif()
