+++ mooncake-integration/store/store_py.cpp
@@ -6,7 +6,7 @@
 #include "dummy_client.h"
 #include "real_client.h"
 #include "types.h"
-#include "p2p_rpc_types.h"
+#include "p2p/master/p2p_rpc_types.h"
 #include "rpc_types.h"
 #include "runtime_config_store.h"
