+++ mooncake-store/benchmarks/CMakeLists.txt
@@ -1,17 +1,17 @@
 # Add allocator benchmark executable
 add_executable(allocator_bench allocator_bench.cpp)
-target_link_libraries(allocator_bench PRIVATE cachelib_memory_allocator mooncake_store)
+target_link_libraries(allocator_bench PRIVATE cachelib_memory_allocator ${MOONCAKE_STORE_LIBS})
 
 # Add master benchmark executable
 add_executable(master_bench master_bench.cpp)
-target_link_libraries(master_bench PRIVATE cachelib_memory_allocator mooncake_store)
+target_link_libraries(master_bench PRIVATE cachelib_memory_allocator ${MOONCAKE_STORE_LIBS})
 
 # Add storage backend benchmark executable
 # This benchmark tests all storage backends (OffsetAllocator, Bucket, FilePerKey)
 # with realistic KV cache workloads for LLM inference scenarios
 add_executable(storage_backend_bench storage_backend_bench.cpp)
 target_link_libraries(storage_backend_bench PRIVATE
-    mooncake_store
+    ${MOONCAKE_STORE_LIBS}
     cachelib_memory_allocator
     gflags::gflags
     glog::glog
