+++ mooncake-store/include/p2p/master/p2p_rpc_service.h
@@ -1,9 +1,9 @@
 #pragma once
 
-#include "p2p_master_service.h"
+#include "p2p/master/p2p_master_service.h"
 #include "rpc_service.h"
 
-#include "p2p_rpc_types.h"
+#include "p2p/master/p2p_rpc_types.h"
 
 namespace mooncake {
