+++ mooncake-store/include/p2p/master/p2p_segment_manager.h
