+++ mooncake-store/include/runtime_config_store.h
@@ -7,7 +7,7 @@
 #include <string>
 #include <variant>
 
-#include "p2p_rpc_types.h"
+#include "p2p/master/p2p_rpc_types.h"
 #include "replica.h"
 #include "rpc_types.h"
 #include "types.h"
