+++ mooncake-store/src/p2p/master/p2p_master_client.cpp
@@ -1,6 +1,6 @@
-#include "p2p_master_client.h"
+#include "p2p/master/p2p_master_client.h"
 
-#include "p2p_rpc_service.h"
+#include "p2p/master/p2p_rpc_service.h"
 #include "utils/scoped_vlog_timer.h"
 
 namespace mooncake {
