+++ mooncake-store/src/p2p/master/p2p_master_metric_manager.cpp
@@ -1,4 +1,4 @@
-#include "p2p_master_metric_manager.h"
+#include "p2p/master/p2p_master_metric_manager.h"
 
 #include <new>
 #include <sstream>  // For string building during serialization
