+++ mooncake-store/src/p2p/master/p2p_segment_manager.cpp
@@ -1,5 +1,5 @@
-#include "p2p_segment_manager.h"
-#include "p2p_master_metric_manager.h"
+#include "p2p/master/p2p_segment_manager.h"
+#include "p2p/master/p2p_master_metric_manager.h"
 #include <glog/logging.h>
 
 namespace mooncake {
