+++ mooncake-store/src/p2p_client_manager.cpp
@@ -1,5 +1,5 @@
 #include "p2p_client_manager.h"
-#include "p2p_master_metric_manager.h"
+#include "p2p/master/p2p_master_metric_manager.h"
 #include "p2p_client_meta.h"
 #include <glog/logging.h>
 #include <algorithm>
