+++ mooncake-store/tests/async_metadata_notifier_test.cpp
@@ -16,7 +16,7 @@
 #define protected public
 #include "async_metadata_notifier.h"
 
-#include "p2p_master_client.h"
+#include "p2p/master/p2p_master_client.h"
 #include "rpc_types.h"
 #include "test_p2p_server_helpers.h"
 #include "types.h"
