+++ mooncake-store/tests/ha/oplog/p2p_hot_standby_service_test.cpp
@@ -16,9 +16,9 @@
 
 #include "ha/oplog/localfs_oplog_store.h"
 #include "mock_oplog_store.h"
-#include "p2p_master_service.h"
-#include "p2p_rpc_service.h"
-#include "p2p_rpc_types.h"
+#include "p2p/master/p2p_master_service.h"
+#include "p2p/master/p2p_rpc_service.h"
+#include "p2p/master/p2p_rpc_types.h"
 
 namespace mooncake::test {
 namespace {
