+++ mooncake-store/tests/p2p_client_meta_test.cpp
@@ -9,7 +9,7 @@
 #include "p2p_client_meta.h"
 #undef private
 #undef protected
-#include "p2p_master_metric_manager.h"
+#include "p2p/master/p2p_master_metric_manager.h"
 
 namespace mooncake {
