+++ mooncake-store/tests/p2p_master_http_endpoints_test.cpp
@@ -30,7 +30,7 @@
 #include <ylt/struct_json/json_reader.h>
 
 #include "master_config.h"
-#include "p2p_rpc_service.h"
+#include "p2p/master/p2p_rpc_service.h"
 #include "types.h"
 #include "utils.h"
