+++ mooncake-store/tests/p2p_master_service_test.cpp
@@ -7,13 +7,13 @@
 
 #define private public
 #define protected public
-#include "p2p_master_service.h"
+#include "p2p/master/p2p_master_service.h"
 #undef protected
 #undef private
 
 #include "master_config.h"
 #include "p2p_client_meta.h"
-#include "p2p_rpc_types.h"
+#include "p2p/master/p2p_rpc_types.h"
 #include "rpc_types.h"
 #include "types.h"
