+++ mooncake-store/tests/test_p2p_server_helpers.h
@@ -10,7 +10,7 @@
 #include <ylt/coro_rpc/coro_rpc_server.hpp>
 
 #include "master_config.h"
-#include "p2p_rpc_service.h"
+#include "p2p/master/p2p_rpc_service.h"
 #include "rpc_service.h"
 #include "types.h"
 #include "utils.h"
